create getSchemaByName and move error checking logic
