+++ src/app/dataValidator/helpers/interpretSchema/getSchemaByName.js
@@ -1,0 +1,12 @@
+const { throwError } = require('../../utilities/messaging');
+const { getAllSchemas } = require('../interpretSchema/getAllSchemas');
+
+const schemas = getAllSchemas();
+
+module.exports.getSchemaByName = schemaName => {
+  if (!(schemaName in schemas)) {
+    throwError(`Error: No schema exists for the block ${schemaName}`);
+  }
+
+  return schemas[schemaName];
+};
