+++ src/app/dataValidator/helpers/validators/validateBlock.js
@@ -1,19 +1,13 @@
-const { log, throwError } = require('../../utilities/messaging');
-const { getAllSchemas } = require('../interpretSchema/getAllSchemas');
+const { log } = require('../../utilities/messaging');
 const { validateNode } = require('./validateNode');
-
-const schemas = getAllSchemas();
+const { getSchemaByName } = require('../interpretSchema/getSchemaByName');
 
 module.exports.validateBlock = dataToValidate => {
   const schemaName = dataToValidate.type;
 
-  if (!(schemaName in schemas)) {
-    throwError(`Error: No schema exists for the block ${dataToValidate.type}`);
-  }
-
   log(`\nValidating block: ${schemaName}`);
   log('----------------------------------------------------------------');
 
-  const blockSchema = schemas[schemaName];
+  const blockSchema = getSchemaByName(schemaName);
   validateNode(blockSchema, dataToValidate, schemaName);
 };
